Update modules.py
Fixing bug in alphafold when recycles=0. (making inference mode consistent with training mode)

We noticed that the behavior is different when setting recycles=0 vs. setting recycles>0 and then modifying num_iter_recycling=0.
When no recycles are defined the prev values are undefined. This disables 3 layers in EmbeddingsAndEvoformer. Even though the inputs to these layers is ZERO, I'm guessing they still contribute a bias to the pair_activations.